test: add on kind integration testing in addition to k3d
Kind is another way of running kubernetes. https://kind.sigs.k8s.io/
Since kind is packaged and released by the CNCF and Kubernetes teams
it tends to be very close to trunk and bleeding edge.

This commit adds kd3 and kind commands to the makefile. Then it wires
the test command to run kind and k3d testing sequentially. This shows
the differences (including the streaming body size issue).