fix: Fix new requets window sizing when window size > request window
Window size is the total number of bytes that can be outstanding.
However Some web socket servers won't accept that in a single inital
frame for a new stream.

Some won't accept more than a frame size. That frame size in in
%Mint.HTTP{server_settings: %{max_frame_size: HERE}} However that's the
maximum that any frame can be. And that includes a whole bunch of
things. So we would preffer to use something more precise.

`dbg` shows that there's a private field on HTTP connections.
request_frame_size that's the default frame size for a new requests
(such as starting a watch or streaming a large PUT). So grab that value
if it's there.